requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
chromadb>=1.0.0
streamlit>=1.30.0
plotly>=5.0.0
